Pasar peliculasSabado/Domindo y accionadoBotonAnadirPanelPeliculas a Modelo
